vioserial: Add speed test to the original serial test
The speed test can be built for both Linux and Windows and intended to run
as following:
* Client in on Windows side.
* Server is on Linux side.
* Client sends a big block prepended by the header.
* Server receives the header, then receives the data, then sends short
respose to the client.
* Client receives the response and can evaluate the speed of the transfer.

On Windows, The test will run the speed test on client mode when the
option "-sp" is used.

When runnig the speed test, the follwing qemu command should be added:
-chardev socket,path=/tmp/foo,id=foo
-device virtserialport,bus=virtio-serial0.0,nr=2,chardev=foo,id=test0,name=test0

Signed-off-by: Ali Abu-Foul <ali@daynix.com>